refactor: simplify transaction list files
