+++ frontend/routes/auth/transactions/home/transactions-list/grouped-transactions-list.spec.tsx
@@ -1,0 +1,7 @@
+describe('GroupedTransactionsList', () => {
+  it('groups all transactions by date', () => {
+    expect(false).toBe(true)
+    // Renders header
+    // Renders list
+  })
+})
