+++ frontend/routes/auth/transactions/home/transactions-list/index.tsx
@@ -1,0 +1,1 @@
+export * from './grouped-transactions-list'
