+++ frontend/routes/auth/transactions/home/transactions-list/transactions-list-empty.tsx
@@ -1,0 +1,11 @@
+export const locators = {
+  transactionListEmpty: 'transaction-list-empty'
+}
+
+export const TransactionListEmpty = () => {
+  return (
+    <div data-testid={locators.transactionListEmpty} className="mt-6 text-sm">
+      No transactions have been placed using this wallet on this network.
+    </div>
+  )
+}
